feat : 유저 정보 받아오기
